Make `hyperref` set pdf metadata
Small thing, but the generated pdf will now correctly list title and
author in its metadata, which is useful if you ever want to import this
into a digital library manager, say.

We also move the loading of `hyperref` to the very end of the preamble
as recommended by the `hyperref` manual.