Fix a bunch of small errors in `p3c10.tex` & general improvements
For the general improvements, we change some `\newcommands` into
operator definitions in the preamble so that the affected commands have
proper spacing. We also remove a superfluous `\te` macro from the
existing operator definitions.

Apart from that, we fix some typos in part 3, chapter 10 ("Duality in
Manifolds"), and most importantly begin messing with some of the
diagrams, replacing scaleboxes with TikZ-internal scaling and proper
font size adjustions. Unfortunately this blows up the formatting on some
pages in a pretty hideous way (since none of the graphics are floated),
so fixing that is now somewhere high up on the TODO-list.