Unify spelling of "neighborhood" & add TODO-list
Some parts of the text, in particular chapter 10 of part 3, contained
the word both spelled as "neighborhood" and "neighbourhood." Adams
himself consistently uses the former, so we follow suit.

We also add a new file, `todo_list.txt`, which (unsurprisingly) contains
a list of things left to-do (right now there's only a single item, but
we'll see).